+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.2
 ----------------------------------------------------------------------
 
+- Update `plot_var_matrix()` to have a `head` argument (0.2.5)
+- Update `plot_time_count_grid()` to have a `min_border_left` argument (0.2.5)
 - Major speedup in `get_subject_data()` (0.2.4)
 - Change default transformation to identity (0.2.3)
 - Handle empty subjects more robustly in fitting all trajectories (0.2.3)